+++ compensation/src/main/java/industrialaccident/domain/CreateCompensationCommand.java
@@ -11,6 +11,6 @@ public class CreateCompensationCommand {
     private Long assessmentId;
     private Long accidentId;
     private String employeeId;
-    private Float averageSalary;
+    private Long averageSalary;
     private Integer period;
 }
